requirements.txt
matplotlib==3.0.2
networkx==2.2
pycorenlp==0.3.0
scikit-learn==0.20.2
scipy==1.2.0
spacy==2.0.18
tensorflow==1.1.0
numpy==1.16.0
# tensorflow-gpu==1.1.0
absl-py
ipython
more_itertools
pandas
plumbum
